+++ src/search.rs
@@ -4,9 +4,24 @@ use strsim::damerau_levenshtein;
 use crate::{
     nodes::Word,
     pages::file_explorer::{FileTree, MdFile},
+    util::CONFIG,
 };
 
 pub fn find_md_files() -> FileTree {
+    let mut ignored_files = Vec::new();
+
+    if CONFIG.gitignore {
+        let gitignore = std::fs::read_to_string(".gitignore");
+        if let Ok(gitignore) = gitignore {
+            for line in gitignore.lines() {
+                if line.starts_with('#') || line.is_empty() {
+                    continue;
+                }
+                ignored_files.push(line.to_string());
+            }
+        }
+    }
+
     let mut tree = FileTree::new();
     let mut stack = vec![std::path::PathBuf::from(".")];
     while let Some(path) = stack.pop() {
@@ -23,9 +38,24 @@ pub fn find_md_files() -> FileTree {
             if path.is_dir() {
                 stack.push(path);
             } else if path.extension().unwrap_or_default() == "md" {
-                let name = path.file_name().unwrap().to_str().unwrap().to_string();
-                let path = path.to_str().unwrap().to_string();
-                tree.add_file(MdFile::new(path, name));
+                let (path_str, path_name) =
+                    if let (Some(path_str), Some(path_name)) = (path.to_str(), path.file_name()) {
+                        (
+                            path_str,
+                            path_name.to_str().unwrap_or("UNKNOWN").to_string(),
+                        )
+                    } else {
+                        continue;
+                    };
+                // Check if the file is in the ignored files list
+                if ignored_files
+                    .iter()
+                    .any(|ignored_file| !find(ignored_file, path_str, 0).is_empty())
+                {
+                    continue;
+                }
+
+                tree.add_file(MdFile::new(path_str.to_string(), path_name));
             }
         }
     }